splitting out actsvg_json module
